refactor(types): remove hub-types and move types into other packges
affects: @esri/hub-common, @esri/hub-sites, @esri/hub-teams
