+++ packages/sites/package.json
@@ -18,8 +18,7 @@
     "@esri/arcgis-rest-request": "^2.13.0",
     "@esri/hub-common": "^6.4.0",
     "@esri/hub-initiatives": "^6.4.0",
-    "@esri/hub-teams": "5.x || 6.x",
-    "@esri/hub-types": "5.x || 6.x"
+    "@esri/hub-teams": "5.x || 6.x"
   },
   "devDependencies": {
     "@esri/arcgis-rest-auth": "^2.14.1",
