+++ packages/sites/src/create-site-model-from-template.ts
@@ -11,7 +11,7 @@ import {
   addSolutionResourceUrlToAssets
 } from "@esri/hub-common";
 import { createHubTeams } from "@esri/hub-teams";
-import { HubTeamType } from "@esri/hub-types";
+import { HubTeamType } from "@esri/hub-teams";
 import { getTheme } from "./get-theme";
 import { getPortalSiteHostname } from "./get-portal-site-hostname";
 import { getPortalSiteUrl } from "./get-portal-site-url";
