+++ packages/sites/src/drafts/apply-draft.ts
@@ -1,6 +1,6 @@
 import { IModel, mergeObjects, cloneObject, getProp } from "@esri/hub-common";
 import { _includeListFromItemType } from "./_include-list-from-item-type";
-import { IDraft } from "@esri/hub-types";
+import { IDraft } from "@esri/hub-common";
 
 /**
  * Applies a draft resource to an item model
