+++ packages/teams/package.json
@@ -17,8 +17,7 @@
     "@esri/arcgis-rest-portal": "^2.15.0",
     "@esri/arcgis-rest-request": "^2.13.0",
     "@esri/arcgis-rest-types": "^2.13.0",
-    "@esri/hub-common": "4.x || 5.x || 6.x",
-    "@esri/hub-types": "5.x || 6.x"
+    "@esri/hub-common": "4.x || 5.x || 6.x"
   },
   "devDependencies": {
     "@esri/arcgis-rest-auth": "^2.14.1",
@@ -26,7 +25,6 @@
     "@esri/arcgis-rest-request": "^2.14.1",
     "@esri/arcgis-rest-types": "^2.15.0",
     "@esri/hub-common": "^6.10.0",
-    "@esri/hub-types": "^6.10.0",
     "@rollup/plugin-commonjs": "^15.0.0",
     "@rollup/plugin-json": "^4.1.0",
     "@rollup/plugin-node-resolve": "^9.0.0",
