+++ packages/teams/src/create-hub-teams.ts
@@ -7,7 +7,7 @@ import {
 } from "@esri/hub-common";
 import { getUserCreatableTeams } from "./utils/get-user-creatable-teams";
 import { _createTeamGroups } from "./utils/_create-team-groups";
-import { HubTeamType } from "@esri/hub-types";
+import { HubTeamType } from "./types";
 import { IGroup } from "@esri/arcgis-rest-types";
 
 /**
