+++ packages/teams/src/index.ts
@@ -1,6 +1,6 @@
 /* Copyright (c) 2018 Environmental Systems Research Institute, Inc.
  * Apache-2.0 */
-
+export * from "./types";
 export * from "./add-users-to-team";
 export * from "./create-hub-team";
 export * from "./create-hub-teams";
