+++ packages/teams/src/types.ts
