+++ packages/teams/src/utils/get-team-properties-map-for-product.ts
@@ -1,4 +1,4 @@
-import { HubProduct } from "@esri/hub-types";
+import { HubProduct } from "../types";
 
 interface ITeamPropMap {
   content: string;
