+++ packages/teams/test/create-hub-teams.test.ts
@@ -3,7 +3,7 @@ import { UserSession } from "@esri/arcgis-rest-auth";
 import { IGroup } from "@esri/arcgis-rest-types";
 import * as _createTeamGroupsModule from "../src/utils/_create-team-groups";
 import { createHubTeams } from "../src/create-hub-teams";
-import { HubTeamType } from "@esri/hub-types";
+import { HubTeamType } from "../src/types";
 
 describe("createHubTeams", () => {
   const ro = {
