+++ packages/teams/test/utils/get-team-types-available-in-product.test.ts
@@ -1,6 +1,6 @@
 import * as getTeamsAvailableInProductModule from "../../src/utils/get-teams-available-in-product";
 import { getTeamTypesAvailableInProduct } from "../../src/utils/get-team-types-available-in-product";
-import { HubTeamType } from "@esri/hub-types";
+import { HubTeamType } from "../../src/types";
 
 describe("getTeamTypesAvailableInProduct", () => {
   it("maps group templates to team types", function() {
