--- packages/types/package.json
@@ -1,58 +1,0 @@
-{
-  "name": "@esri/hub-types",
-  "version": "6.10.0",
-  "description": "Shared Typescript types for @esri/hub.js",
-  "main": "dist/node/index.js",
-  "unpkg": "dist/umd/types.umd.js",
-  "module": "dist/esm/index.js",
-  "js:next": "dist/esm/index.js",
-  "sideEffects": false,
-  "types": "dist/esm/index.d.ts",
-  "license": "Apache-2.0",
-  "dependencies": {
-    "tslib": "^1.13.0"
-  },
-  "peerDependencies": {
-    "@esri/arcgis-rest-types": "^2.13.0"
-  },
-  "devDependencies": {
-    "@esri/arcgis-rest-types": "^2.15.0",
-    "@rollup/plugin-commonjs": "^15.0.0",
-    "@rollup/plugin-json": "^4.1.0",
-    "@rollup/plugin-node-resolve": "^9.0.0",
-    "rollup": "^2.26.5",
-    "rollup-plugin-filesize": "^9.0.2",
-    "rollup-plugin-typescript2": "^0.27.2",
-    "rollup-plugin-uglify": "^6.0.4",
-    "typescript": "^3.8.1"
-  },
-  "files": [
-    "dist/**"
-  ],
-  "scripts": {
-    "prepare": "npm run build",
-    "build": "npm run build:node && npm run build:umd && npm run build:esm",
-    "build:esm": "tsc --module es2015 --outDir ./dist/esm --declaration",
-    "build:umd": "rollup -c ../../umd-base-profile.js && rollup -c ../../umd-production-profile.js",
-    "build:node": "tsc --module commonjs --outDir ./dist/node",
-    "dev:esm": "tsc -w --module es2015 --outDir ./dist/esm --declaration",
-    "dev:umd": "rollup -w -c ../../umd-base-profile.js",
-    "dev:node": "tsc -w --module commonjs --outDir ./dist/node",
-    "format": "prettier --write ./{src,test}/**/*.ts",
-    "format:check": "prettier --check ./{src,test}/**/*.ts"
-  },
-  "publishConfig": {
-    "access": "public"
-  },
-  "repository": "github:Esri/hub.js",
-  "contributors": [
-    {
-      "name": "Dave Bouwman",
-      "email": "dbouwman@esri.com"
-    }
-  ],
-  "bugs": {
-    "url": "https://github.com/Esri/hub.js/issues"
-  },
-  "homepage": "https://github.com/Esri/hub.js#readme"
-}
