--- packages/types/src/index.ts
@@ -1,5 +1,0 @@
-/* Copyright (c) 2018 Environmental Systems Research Institute, Inc.
- * Apache-2.0 */
-
-export * from "./models";
-export * from "./team-types";
