--- packages/types/src/models.ts
@@ -1,21 +1,0 @@
-/* Copyright (c) 2018 Environmental Systems Research Institute, Inc.
- * Apache-2.0 */
-
-import { IItem } from "@esri/arcgis-rest-types";
-
-/**
- * Generic Model, used with all items that have a json
- * `/data` payload
- *
- * @export
- * @interface IModel
- */
-export interface IModel {
-  item: IItem;
-  data?: {
-    [propName: string]: any;
-  };
-  [key: string]: any;
-}
-
-export type IDraft = Partial<IModel>;
